requirements.txt
asgiref==3.7.2
cffi==1.15.1
contextlib2==21.6.0
cryptography==41.0.1
cycler==0.11.0
distlib==0.3.6
Django~=4.2.2
django-cors-headers==4.0.0
django-ninja==0.22.1
django-ninja-extra==0.18.9
django-ninja-jwt==5.2.5
filelock==3.8.0
fonttools==4.34.4
injector==0.20.1
kiwisolver==1.4.4
networkx==2.8.5
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil~=2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.64.0
typing_extensions==4.6.3
virtualenv==20.16.3
orjson~=3.9.1
